fixup! refactor: check all possible conditions before moving
